[nix] Use Home Manager flake for nixpkgs
This sets up the local Flake Registry so that a bare "nixpkgs" refers to
the same revision of the nixpkgs flake that was used to build the
current home-manager environment. This should increase the cache hit
rate when, for example, setting up a temporary environment with `nix
shell nixpkgs#foo`.